refactor(mui): move crud hooks to CrudPanelView
