--- libs/mui/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./useCrudModalForm";
-export * from "./useCrudTableItemAction";
