+++ libs/mui/src/index.ts
@@ -1,4 +1,3 @@
 export * from './@types';
 export * from './components';
-export * from './hooks';
 export * from './views';
