+++ libs/mui/src/views/CrudPanelView/hooks/useCrudTableItemAction/props.ts
@@ -12,7 +12,7 @@ import {
 import { ButtonProps } from '@mui/material/Button';
 import { TooltipProps } from '@mui/material/Tooltip';
 
-import { DialogProps, DialogRefProps } from '../../components/Dialog';
+import { DialogProps, DialogRefProps } from '../../../../components/Dialog';
 
 /**
  * ===========================
@@ -158,3 +158,10 @@ export type CrudTableItemActionProps<
     CrudGraphApiExportType<TSchema>
   >['onClick'];
 };
+
+/**
+ * ===========================
+ * EXPORTS
+ * ===========================
+ */
+export default CrudTableItemActionProps;
